+++ src/forms/FormRadio/FormRadio.fixture.jsx
@@ -22,6 +22,9 @@ class DemoFormRadio extends React.PureComponent /** @lends @FormRadio.prototype 
   handleChange = ({ /* id, */ value }) => {
     this.setState({ value })
   }
+  setRef = (node) => {
+    node && typeof node.focus === 'function' && node.focus()
+  }
   render() {
     const { value } = this.state
     return (
@@ -29,14 +32,14 @@ class DemoFormRadio extends React.PureComponent /** @lends @FormRadio.prototype 
         id="DemoRadio"
         onChange={this.handleChange}
         value={value}
+        ref={this.setRef}
       />
     )
   }
 }
 
-
 export default {
-  demo: <DemoFormRadio />,
-  // default: <FormRadio className="extraButtonClass" />,
-  // checked: <FormRadio value={true} />,
+  default: <FormRadio className="extraButtonClass" />,
+  primaryChecked: <FormRadio value={true} theme="primary" />,
+  demoWithState: <DemoFormRadio />,
 }